Tests/LibWeb/Text/expected/canvas/webgl-oes-vertex-array-object-binding.txt

binding before enabling extension: null
error: INVALID_ENUM
initial binding: null
bound VAO is returned: true
binding after unbind: null
binding after deleting bound VAO: null
error: NO_ERROR
